:clapper: update example
